requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
bleach==4.1.0
bokeh==2.3.3
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
einops==0.4.1
entrypoints==0.4
faiss==1.7.2
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.0.8
idna==3.3
idna-ssl==1.1.0
immutables==0.17
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonlines==3.0.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.1
lmdb==1.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
more-itertools==8.12.0
multidict==5.2.0
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.0
Pygments==2.12.0
pyparsing==3.0.8
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.3.3
pytrec-eval==0.5
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sqlitedict==2.0.0
sqlitelist==0.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.4.0
torchmetrics==0.2.0
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
transformers==4.6.1
typed-argument-parser==1.7.2
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
xeus-python==0.9.1
yarl==1.7.2
zipp==3.6.0
